Pass red states as method parameters.
Also, don't pass l1. Instead, initialize it lazily. Fun fact: there was
a net savings of 29 lines by virtualizing l1 and only instantiating it
immediately before usage! That will be the only savings, though.

Inline most of the parsing methods. This doesn't yet open up their
preconditions, but it's progress. It sure did make this commit larger!

Add a line buffer `lb` which holds all emitted text. It is
state-threaded just like the other reds. Eventually we may use something
like rlib.rope for this if necessary, but I really value persistence
here over the requirement to add auxiliary methods for managing e.g. a
list of lines. (That might be an interesting thought to investigate
later on when generalizing output...)

Clean up the preconditions on the loops. There's now a clear need for a
.fork and .join semantics!