Use a stack for the margin.
This allows us to push the current margin before starting a rule that
might push an unbalanced margin when recursing; at the end of the loop,
we can pop the margin to restore it.

The resulting code is a little funky in the control flow? But maybe
that's just my intuition.