More wiring up of custom output.
Need to sleep. Making progress though.

This is kind of a big switchover, but I have confidence in the overall
structure. The real fun is going to be in fixing the translation errors.

I'm not cutting over yet, though. RPython expects all of the types to
unify and I may need to add a Box-like abstraction.