Remove a few unused features.
The idea is that we'll tighten up to a minimal syntax, self-host the
newer semantic-PEG stuff, and then add in whatever still makes sense.

Also, add the length-of-register primitive operation to output, because
it carries its own weight.

I think we're ready. At this point, we're going to see what happens when
we remove the old rules and run entirely on the new Z rules.