Big leap!
So far: That wasn't the worst parser to hand-write TBH. Also it was much
shorter than I had feared.

We are able to translate and parse two iterations. The bootstrap is 125
lines so far.

The third iteration is miscompiled. Not sure exactly when things go
wrong, but we must fix it.