Set up functors.
Not sure of the correct order for parse rules. I vaguely feel like I
want to introduce rules before using them, and order them in the order
that a document will use them in, but those are vague urges rather than
technical requirements.

I think that we're having a limited success! However, there is a slight
problem: the generated functors seem to degenerate when iterated. The
peg2 functor is fine, but the production functor is not. The root issue
was that this fragment:

    "'" + s + "'"

Is not interchangeable when the wrong quotes are used:

    """ + s + """

Thanks to Python's triple-quotes. With that handled, we are now able to
generate the functors.