WIP different approach to rules
This approach is basically an attempt at ACE-matching. We have one
primitive type of rule: match a single pattern and add a representative
to its e-class, preferring the new representative over others.

The syntax supports A-matching by allowing a special functor
builtin.List/* of arbitrary arity. Any particular builtin.List has
definite finite arity, but the patterns for lists can match any arity
(above an easy-to-compute minimum arity, if needed as a later
optimization!) As a consequence, peephole optimization appears to be
expressible!