bf: Use character classes.
This lengthens the parser but makes it more efficient. It also
increases the elegance of the generated code by removing whitespace as a
special case and not using the token-handling shortcut, which is nice.

I'm wondering about a feature akin to that in the Yacc lineage where a
character class can be designated as the "ignore" or "whitespace" class.
The obvious issue with this is that languages with strings immediately
need to add some sort of stateful lexer, which quickly causes the parser
to break apart into regular and context-free pieces. While I'm not
against regular expressions, I do not like that sort of lexing.