WIP make objects on union-find heap
We had to get rid of production.Tuple/1, but everything else is strongly
typed. Concatenation has been replaced with flattening for lists so that
the remanining production.PlusMany/1 operation is only applicable to
strings. This will allow us to replace strings with ints throughout the
interface.

It looks like building structures will not be hard. Extracting them will
be tricky though. I hand-wrote an extractor for builtins in order to
understand what can be done, and it's not looking great.

I think that we can build a separate functor for generating the
extractions. This could also lead to a new scheme for compiling rules,
so that we don't have to write those out by hand either.

...But I might just take a big leap. Not sure!