WIP write down desired rules
This is the final bit of dogfooding that we need. It's going well so
far, although I'm noticing a fair amount of circularity and redundancy
so far.

I can't promise that the functors will survive. I might get rid of them
or relegate them to defining output formats only.

We added syntax for let-statements in rules which create constants. The
syntax appears to work. Also, those let-statements are beginning to
exercise the relational storage.