Swim to shore.
We're almost there, I think. I was worried about how to generate the
multi-relation namespacing classes, but I think that the trick of
pre-recording the names will work.

We have two translations in a row! The third translation successfully
parses, but doesn't produce a reducible tree for some reason. I think
that the reason is that we are calling findStr on a py.Statement; we
either need to be lazier or emit cleaner code.

Trying the Flatten strategy on the other three ops that could be a
problem: Length, Concat, and Join. Join works? Length works? Concat
works! There were many places where I fucked up the big towers of commas
and brackets, but fixing those, we've generated something that looks
like it might self-compile. So I think we're gonna start iterating on
that now.

A single beEmpty is missing and there are segfaults. I'm not happy with
this yet. But we're almost there.